fix(tests): eliminate catalog-data fixture race in CI
Move the catalog-data fixture bootstrap out of per-test-file
ensureTestDirs() into a vitest globalSetup so it runs exactly once
before any test file. This eliminates the parallel ENOENT/EEXIST race
that occurs when multiple vitest workers simultaneously rm/mkdir/cp
the shared catalog-data directory.

Extract the bootstrap logic into a shared helper
(tests/helpers/catalogDataFixture.ts) used by both the vitest and
Playwright global setups, replacing duplicated inline copies.
