+++ tests/integration/helpers/globalSetup.ts
@@ -1,0 +1,5 @@
+import { ensureCatalogFixture } from '../../helpers/catalogDataFixture.js';
+
+export default async function globalSetup() {
+  await ensureCatalogFixture();
+}
