+++ vitest.config.ts
@@ -4,6 +4,7 @@ import path from 'path';
 export default defineConfig({
   test: {
     environment: 'node',
+    globalSetup: ['tests/integration/helpers/globalSetup.ts'],
     include: ['tests/**/*.test.ts'],
     exclude: ['tests/e2e/**'],
     coverage: {
